+++ app/src/main/java/com/example/musinsa/adapters/CustomRecyclerViewAdapter.kt
@@ -7,7 +7,6 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.musinsa.databinding.*
 import com.example.musinsa.model.Item
-import java.lang.IllegalArgumentException
 
 class CustomRecyclerViewAdapter :
     ListAdapter<Item.ItemType, RecyclerView.ViewHolder>(ItemDiffUtil) {
@@ -20,13 +19,13 @@ class CustomRecyclerViewAdapter :
             HEADER -> HeaderViewHolder(
                 ItemRecyclerHeaderBinding.inflate(inflater, parent, false)
             )
-            CONTENTS_GOODS -> ItemViewHolder(
+            CONTENTS_GOODS -> ItemGoodsViewHolder(
                 ItemRecyclerGoodsBinding.inflate(inflater, parent, false)
             )
             CONTENTS_STYLE -> ItemStyleHolder(
                 ItemRecyclerStyleBinding.inflate(inflater, parent, false)
             )
-            CONTENTS_BANNER -> BannerViewHolder(
+            CONTENTS_BANNER -> ItemBannerViewHolder(
                 ItemViewpagerBinding.inflate(inflater, parent, false)
             )
             FOOTER -> FooterViewHolder(
@@ -43,9 +42,9 @@ class CustomRecyclerViewAdapter :
         val item = getItem(position)
         when (holder) {
             is HeaderViewHolder -> holder.bind(item as Item.ItemType.Header)
-            is ItemViewHolder -> holder.bind(item as Item.ItemType.Contents.Goods)
+            is ItemGoodsViewHolder -> holder.bind(item as Item.ItemType.Contents.Goods)
             is ItemStyleHolder -> holder.bind(item as Item.ItemType.Contents.Style)
-            is BannerViewHolder -> holder.bind(item as Item.ItemType.Contents.Banner)
+            is ItemBannerViewHolder -> holder.bind(item as Item.ItemType.Contents.Banner)
             is FooterViewHolder -> holder.bind(item as Item.ItemType.Footer)
         }
     }
@@ -60,11 +59,27 @@ class CustomRecyclerViewAdapter :
         }
     }
 
-    class ItemViewHolder(
+    class ItemBannerViewHolder(
+        private val binding: ItemViewpagerBinding,
+    ) : RecyclerView.ViewHolder(binding.root) {
+        fun bind(item: Item.ItemType.Contents.Banner) {
+            binding.item = item
+        }
+    }
+
+    class ItemGoodsViewHolder(
         private val binding: ItemRecyclerGoodsBinding,
     ) : RecyclerView.ViewHolder(binding.root) {
         fun bind(item: Item.ItemType.Contents.Goods) {
             binding.item = item
+        }
+    }
+
+    class ItemStyleHolder(
+        private val binding: ItemRecyclerStyleBinding,
+    ) : RecyclerView.ViewHolder(binding.root) {
+        fun bind(item: Item.ItemType.Contents.Style) {
+            binding.style = item
         }
     }
 
@@ -81,22 +96,6 @@ class CustomRecyclerViewAdapter :
     ) : RecyclerView.ViewHolder(binding.root) {
         fun bind(item: Item.ItemType.Footer) {
             binding.isRefresh = true
-        }
-    }
-
-    class ItemStyleHolder(
-        private val binding: ItemRecyclerStyleBinding,
-    ) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: Item.ItemType.Contents.Style) {
-            binding.style = item
-        }
-    }
-
-    class BannerViewHolder(
-        private val binding: ItemViewpagerBinding,
-    ) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: Item.ItemType.Contents.Banner) {
-            binding.item = item
         }
     }
