+++ app/src/main/java/com/example/musinsa/datasource/DataSourceImpl.kt
@@ -8,5 +8,6 @@ import javax.inject.Singleton
 class DataSourceImpl @Inject constructor(
     private val apiClient: ApiClient,
 ) : DataSource {
+
     override suspend fun getItemList() = apiClient.getItemList()
 }
