+++ app/src/main/java/com/example/musinsa/ui/MainActivity.kt
@@ -1,7 +1,7 @@
 package com.example.musinsa.ui
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import com.example.musinsa.R
 import com.example.musinsa.databinding.ActivityMainBinding
