+++ app/src/main/res/layout/item_viewpager.xml
@@ -19,10 +19,10 @@
             imgUrl="@{item.thumbnailUrl}"
             android:layout_width="match_parent"
             android:layout_height="400dp"
+            android:contentDescription="@string/banner_area"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            android:contentDescription="@string/banner_area" />
+            app:layout_constraintTop_toTopOf="parent" />
 
         <TextView
             android:id="@+id/tv_title"
